+++ examples/04-logger/app/modules/winston/winston.module.ts
@@ -47,7 +47,8 @@ export class WinstonModule {
       transports: [transport],
     });
 
-    (logger as any).setLevel = (value: LogLevels) => {
+    // Logger must have `setLevel` method.
+    (logger as unknown as Logger).setLevel = (value: LogLevels) => {
       logger.level = value;
     };